Use getPkgVersion
