+++ packages/@expo/metro-config/src/babel-transformer.ts
@@ -9,10 +9,13 @@
 // and adds support for web and Node.js environments via `isServer` on the Babel caller.
 import type { BabelTransformer, BabelTransformerArgs } from '@expo/metro/metro-babel-transformer';
 import assert from 'node:assert';
+import path from 'path';
+import resolveFrom from 'resolve-from';
 
 import type { TransformOptions } from './babel-core';
 import { loadBabelConfig } from './loadBabelConfig';
 import { transformSync } from './transformSync';
+import { getPkgVersion } from './utils/getPkgVersion';
 
 export type ExpoBabelCaller = TransformOptions['caller'] & {
   babelRuntimeVersion?: string;
@@ -59,13 +62,12 @@ const memoizeWarning = memoize((message: string) => {
   debug(message);
 });
 
-function getIsHermesV1(): boolean {
-  try {
-    const { version } = require('hermes-compiler/package.json');
-    return typeof version === 'string' && version.startsWith('250829098');
-  } catch {
-    return false;
-  }
+function getIsHermesV1(projectRoot: string): boolean {
+  const reactNativePath = resolveFrom.silent(projectRoot, 'react-native/package.json');
+  if (!reactNativePath) return false;
+
+  const hermesVersion = getPkgVersion(path.dirname(reactNativePath), 'hermes-compiler');
+  return typeof hermesVersion === 'string' && hermesVersion.startsWith('250829098');
 }
 
 function getBabelCaller({
@@ -121,7 +123,7 @@ function getBabelCaller({
     // target environment.
     engine: stringOrUndefined(options.customTransformOptions?.engine),
     // Indicate whether the project is using Hermes V1 (hermes-compiler version 250829098.x).
-    isHermesV1: getIsHermesV1(),
+    isHermesV1: getIsHermesV1(options.projectRoot),
 
     // Provide the project root for accurately reading the Expo config.
     projectRoot: options.projectRoot,
